+++ src/app/core/index.ts
@@ -1,2 +1,3 @@
 export * from './guards/auth.guard';
 export * from './services/feedbacks.service';
+export * from './decorators';