+++ src/app/products/models/category.enum.ts
@@ -1,7 +1,7 @@
 export enum ProductCategory {
-  Category1 = 'category1',
-  Category2 = 'category2',
-  Category3 = 'category3',
-  Category4 = 'category4',
-  Category5 = 'category5'
+  Category1,
+  Category2,
+  Category3,
+  Category4,
+  Category5
 }